build: estimate gas cost of deployment
- store in folder uploaded as circle artifact
- deploy script converted to typescript
